third_party/blink/web_tests/fast/js/global-function-resolve-expected.txt

Test to make sure cached lookups on the global object are performed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cachedFunctions[0]() is addEventListener
PASS cachedFunctions[0]() is addEventListener
PASS cachedFunctions[1]() is alert
PASS cachedFunctions[1]() is alert
PASS cachedFunctions[2]() is captureEvents
PASS cachedFunctions[2]() is captureEvents
PASS cachedFunctions[3]() is clearInterval
PASS cachedFunctions[3]() is clearInterval
PASS cachedFunctions[4]() is clearTimeout
PASS cachedFunctions[4]() is clearTimeout
PASS cachedFunctions[5]() is clientInformation
PASS cachedFunctions[5]() is clientInformation
PASS cachedFunctions[6]() is close
PASS cachedFunctions[6]() is close
PASS cachedFunctions[7]() is closed
PASS cachedFunctions[7]() is closed
PASS cachedFunctions[8]() is confirm
PASS cachedFunctions[8]() is confirm
PASS cachedFunctions[9]() is console
PASS cachedFunctions[9]() is console
PASS cachedFunctions[10]() is crypto
PASS cachedFunctions[10]() is crypto
PASS cachedFunctions[11]() is defaultStatus
PASS cachedFunctions[11]() is defaultStatus
PASS cachedFunctions[12]() is defaultstatus
PASS cachedFunctions[12]() is defaultstatus
PASS cachedFunctions[13]() is description
PASS cachedFunctions[13]() is description
PASS cachedFunctions[14]() is devicePixelRatio
PASS cachedFunctions[14]() is devicePixelRatio
PASS cachedFunctions[15]() is dispatchEvent
PASS cachedFunctions[15]() is dispatchEvent
PASS cachedFunctions[16]() is document
PASS cachedFunctions[16]() is document
PASS cachedFunctions[17]() is getComputedStyle
PASS cachedFunctions[17]() is getComputedStyle
PASS cachedFunctions[18]() is getSelection
PASS cachedFunctions[18]() is getSelection
PASS cachedFunctions[19]() is history
PASS cachedFunctions[19]() is history
PASS cachedFunctions[20]() is innerHeight
PASS cachedFunctions[20]() is innerHeight
PASS cachedFunctions[21]() is innerWidth
PASS cachedFunctions[21]() is innerWidth
PASS cachedFunctions[22]() is location
PASS cachedFunctions[22]() is location
PASS cachedFunctions[23]() is locationbar
PASS cachedFunctions[23]() is locationbar
PASS cachedFunctions[24]() is menubar
PASS cachedFunctions[24]() is menubar
PASS cachedFunctions[25]() is moveBy
PASS cachedFunctions[25]() is moveBy
PASS cachedFunctions[26]() is moveTo
PASS cachedFunctions[26]() is moveTo
PASS cachedFunctions[27]() is name
PASS cachedFunctions[27]() is name
PASS cachedFunctions[28]() is navigator
PASS cachedFunctions[28]() is navigator
PASS cachedFunctions[29]() is open
PASS cachedFunctions[29]() is open
PASS cachedFunctions[30]() is openDatabase
PASS cachedFunctions[30]() is openDatabase
PASS cachedFunctions[31]() is opener
PASS cachedFunctions[31]() is opener
PASS cachedFunctions[32]() is outerHeight
PASS cachedFunctions[32]() is outerHeight
PASS cachedFunctions[33]() is outerWidth
PASS cachedFunctions[33]() is outerWidth
PASS cachedFunctions[34]() is pageXOffset
PASS cachedFunctions[34]() is pageXOffset
PASS cachedFunctions[35]() is pageYOffset
PASS cachedFunctions[35]() is pageYOffset
PASS cachedFunctions[36]() is parent
PASS cachedFunctions[36]() is parent
PASS cachedFunctions[37]() is prompt
PASS cachedFunctions[37]() is prompt
PASS cachedFunctions[38]() is releaseEvents
PASS cachedFunctions[38]() is releaseEvents
PASS cachedFunctions[39]() is removeEventListener
PASS cachedFunctions[39]() is removeEventListener
PASS cachedFunctions[40]() is resizeBy
PASS cachedFunctions[40]() is resizeBy
PASS cachedFunctions[41]() is resizeTo
PASS cachedFunctions[41]() is resizeTo
PASS cachedFunctions[42]() is screen
PASS cachedFunctions[42]() is screen
PASS cachedFunctions[43]() is screenLeft
PASS cachedFunctions[43]() is screenLeft
PASS cachedFunctions[44]() is screenTop
PASS cachedFunctions[44]() is screenTop
PASS cachedFunctions[45]() is screenX
PASS cachedFunctions[45]() is screenX
PASS cachedFunctions[46]() is screenY
PASS cachedFunctions[46]() is screenY
PASS cachedFunctions[47]() is scroll
PASS cachedFunctions[47]() is scroll
PASS cachedFunctions[48]() is scrollBy
PASS cachedFunctions[48]() is scrollBy
PASS cachedFunctions[49]() is scrollTo
PASS cachedFunctions[49]() is scrollTo
PASS cachedFunctions[50]() is scrollX
PASS cachedFunctions[50]() is scrollX
PASS cachedFunctions[51]() is scrollY
PASS cachedFunctions[51]() is scrollY
PASS cachedFunctions[52]() is setInterval
PASS cachedFunctions[52]() is setInterval
PASS cachedFunctions[53]() is setTimeout
PASS cachedFunctions[53]() is setTimeout
PASS cachedFunctions[54]() is status
PASS cachedFunctions[54]() is status
PASS cachedFunctions[55]() is stop
PASS cachedFunctions[55]() is stop
PASS cachedFunctions[56]() is window
PASS cachedFunctions[56]() is window
PASS successfullyParsed is true

TEST COMPLETE
